requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
google-generativeai>=0.5.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
python-dateutil
aiofiles
httpx
numpy
python-multipart
duckdb
markdown
beautifulsoup4
pillow
requests
pytest